last check

+++ ft_bzero.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_bzero.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: kayyilma <kayyilma@student.42kocaeli.co    +#+  +:+       +#+        */
+/*   By: kayyilma <kayyilma@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/03 15:45:24 by kayyilma          #+#    #+#             */
-/*   Updated: 2022/10/16 01:56:23 by kayyilma         ###   ########.fr       */
+/*   Updated: 2022/11/21 18:08:53 by kayyilma         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 void	bzero(void *s, size_t n)
 {
-	int		index;
-	char	*tmp;
+	size_t		index;
+	char		*tmp;
 
 	index = 0;
 	tmp = (char *)s;
